apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==1.10.15
python-multipart==0.0.9
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.3.0
httpx==0.27.0
# readability-lxml and lxml were removed because they require C extensions on macOS